# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BPS looker read API",
    "description": "Read-only access to the crawled BPS data (labeled `v_*` views).",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Liveness + database reachability",
        "description": "200 with the DB's Alembic revision, or 503 when the database can't be reached.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/domains": {
      "get": {
        "tags": [
          "domains"
        ],
        "summary": "BPS domains (national, provinces, regencies)",
        "description": "Every crawled domain, ordered by ``domain_id``; ``level`` filters pusat/prov/kab.",
        "operationId": "listDomains",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pusat",
                    "prov",
                    "kab"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only domains at this level.",
              "title": "Level"
            },
            "description": "Only domains at this level."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Domain"
                  },
                  "title": "Response Listdomains"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/variables": {
      "get": {
        "tags": [
          "variables"
        ],
        "summary": "Search the dynamic-table variable catalog",
        "description": "Full-text search ranked by relevance (title over subject), or all by title when ``q`` is\nempty; ``total`` counts every match.",
        "operationId": "searchVariables",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Words to find in the title or subject (all must match, each as a prefix: `infl bul`). Empty → every variable, by title.",
              "title": "Q"
            },
            "description": "Words to find in the title or subject (all must match, each as a prefix: `infl bul`). Empty → every variable, by title."
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this domain (e.g. `0000`).",
              "title": "Domain"
            },
            "description": "Only this domain (e.g. `0000`)."
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pusat",
                    "prov",
                    "kab"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only domains at this level.",
              "title": "Level"
            },
            "description": "Only domains at this level."
          },
          {
            "name": "subject",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this subject (`subject_id`).",
              "title": "Subject"
            },
            "description": "Only this subject (`subject_id`)."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariablePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Domain": {
        "properties": {
          "domain_id": {
            "type": "string",
            "title": "Domain Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "level": {
            "type": "string",
            "enum": [
              "pusat",
              "prov",
              "kab"
            ],
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "domain_id",
          "name",
          "level"
        ],
        "title": "Domain"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status"
          },
          "database": {
            "type": "string",
            "enum": [
              "ok",
              "unreachable"
            ],
            "title": "Database"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "status",
          "database"
        ],
        "title": "Health"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VariablePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VariableSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "VariablePage"
      },
      "VariableSummary": {
        "properties": {
          "domain_id": {
            "type": "string",
            "title": "Domain Id"
          },
          "domain_name": {
            "type": "string",
            "title": "Domain Name"
          },
          "domain_level": {
            "type": "string",
            "enum": [
              "pusat",
              "prov",
              "kab"
            ],
            "title": "Domain Level"
          },
          "var_id": {
            "type": "integer",
            "title": "Var Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "domain_id",
          "domain_name",
          "domain_level",
          "var_id",
          "title"
        ],
        "title": "VariableSummary",
        "description": "A dynamic-table variable as listed by search (no definition/notes — see U3 detail)."
      }
    }
  },
  "tags": [
    {
      "name": "meta",
      "description": "Service health."
    },
    {
      "name": "domains",
      "description": "BPS domains: national (pusat), provinces, regencies."
    },
    {
      "name": "variables",
      "description": "Dynamic-table variables: search, metadata, data."
    }
  ]
}